refactor(client): fix imports
