+++ packages/client/src/components/AppNavBar/index.ts
@@ -1,0 +1,1 @@
+export * from './AppNavBar';
