+++ packages/client/src/components/Hero/index.ts
@@ -1,0 +1,1 @@
+export * from './Hero';
