+++ packages/client/src/components/HomeSteps/index.ts
@@ -1,0 +1,1 @@
+export * from './HomeSteps';
