+++ packages/client/src/components/MediaCard/index.ts
@@ -1,0 +1,2 @@
+export * from './MediaCard';
+export * from './MediaCardsGrid';
