+++ packages/client/src/components/index.ts
@@ -1,7 +1,5 @@
 export * from './AppNavBar';
 export * from './Footer';
 export * from './Hero';
-export * from './HomeStep';
 export * from './HomeSteps';
 export * from './MediaCard';
-export * from './MediaCardsGrid';
