+++ packages/client/src/config/index.ts
@@ -1,0 +1,2 @@
+export * from './routes';
+export * from './theme';
