+++ packages/client/src/main.tsx
@@ -1,10 +1,10 @@
-import './index.css';
+import './config/index.css';
 
 import React from 'react';
 import ReactDOM from 'react-dom';
 
 import { App } from './App';
-import { ThemeContextProvider } from './context';
+import { ThemeContextProvider } from './contexts';
 
 ReactDOM.render(
   <React.StrictMode>
